This and ServerAndApp classes outside Application.h|cpp files
