+++ cpp/api/include/Defines.h
@@ -35,5 +35,16 @@
   #endif
 #endif
 
+// Using Visual Studio preprocessor.
+// It must be improved in case of other compilers.
+#ifdef _WIN32
+	#include <process.h>
+	#define GET_PROCESS_PID() _getpid()
+#else
+	#include <unistd.h>
+
+	#define GET_PROCESS_PID() ::getpid()
+#endif
+
 
 #endif
