+++ cpp/api/src/base/Waiting.cpp
@@ -17,7 +17,7 @@
 #include "Waiting.h"
 #include "WaitingSet.h"
 #include "Application.h"
-
+#include "This.h"
 #include <iostream>
 
 using namespace std;
