+++ cpp/api/src/coms/BasicResponder.cpp
@@ -16,6 +16,7 @@
 
 #include "BasicResponder.h"
 
+#include "This.h"
 #include "ImplFactory.h"
 #include "RequestSocket.h"
 #include "Messages.h"
