+++ cpp/api/src/coms/MultiResponder.cpp
@@ -15,6 +15,8 @@
  */
 
 #include "MultiResponder.h"
+
+#include "This.h"
 #include "ImplFactory.h"
 #include "RequestSocket.h"
 #include "Messages.h"
